refactor(cli): add file extension on import paths
